Adding new swift files to allow modularity. Changed functionality of buttons in first main screen to send data to next view. Testing if values are passed into said next view.
