+++ modulo/SimpleCalcTipChoiceView.swift
@@ -7,50 +7,20 @@
 
 import SwiftUI
 
-struct TipCalculationValues {
-    
-    var totalCost: Double
-    var tipPercentage: Int
-    
-    func calculateTipPercentageDivided() -> Double {
-        let tipPercentageDivided = Double(tipPercentage) / 100
-        return tipPercentageDivided
-    }
-    
-    func calculateTipAmount() -> Double {
-        let tipAmount = totalCost * calculateTipPercentageDivided()
-        return tipAmount
-    }
-    
-    func calculateTipAmountRounded() -> Double {
-        let roundedTipAmount = (calculateTipAmount() * 100).rounded() / 100
-        return roundedTipAmount
-    }
-    
-    func calculateTotalCostInclTipAmt() -> Double {
-        let totalCostInclTipAmt = totalCost + calculateTipAmount()
-        return totalCostInclTipAmt
-    }
-    
-    func calculateTotalAmtInclTipRounded() -> Double {
-        let roundedTotalCostInclTipAmt = (calculateTotalCostInclTipAmt() * 100).rounded() / 100
-        return roundedTotalCostInclTipAmt
-    }
-    
-}
-
 struct SimpleCalcTipChoiceView: View {
 
+    @ObservedObject var viewModel: RatingViewModel
     
     var body: some View {
         
         ZStack {
             
-            Color(red: 0.0, green: 0.0, blue: 1.0, opacity: 1.0)
-                .ignoresSafeArea()
+            let bingBong = viewModel.selectedRating ?? 0
             
-            Text("TipChoice")
-                .foregroundColor(.white)
+            Text(String(bingBong))
+                        
+            //Color(red: 0.0, green: 0.0, blue: 1.0, opacity: 1.0)
+                //.ignoresSafeArea()
             
         }
         .navigationBarHidden(true)
@@ -62,6 +32,8 @@ struct SimpleCalcTipChoiceView: View {
 
 struct SimpleCalcTipChoiceView_Previews: PreviewProvider {
     static var previews: some View {
-        SimpleCalcTipChoiceView()
+        let dummyViewModel = RatingViewModel()
+
+        SimpleCalcTipChoiceView(viewModel: dummyViewModel)
     }
 }
